Fix build after move of ThreadPool to shared partition.
