+++ src/System.Private.CoreLib/src/Resources/Strings.resx
@@ -732,6 +732,9 @@
   <data name="ArgumentOutOfRange_LengthTooLarge" xml:space="preserve">
     <value>The specified length exceeds the maximum value of {0}.</value>
   </data>
+  <data name="ArgumentOutOfRange_LessEqualToIntegerMaxVal" xml:space="preserve">
+    <value>Argument must be less than or equal to 2^31 - 1 milliseconds.</value>
+  </data>
   <data name="ArgumentOutOfRange_NeedNonNegNum" xml:space="preserve">
     <value>Non-negative number required.</value>
   </data>
